move all files to .ts to prepare for a merge
