+++ Model/Quotes.ts
