+++ cogs/damnboi.ts
