+++ cogs/quotes.ts
